+++ base/src/user_commands/signed_command/builder.rs
@@ -3,12 +3,11 @@
 
 //! Helpers for building a user command
 
-use proof_systems::mina_signer::{CompressedPubKey, Keypair, NetworkId, Signer};
+use proof_systems::mina_signer::CompressedPubKey;
 
 use crate::numbers::{AccountNonce, Amount, GlobalSlotNumber, TokenId};
 use crate::user_commands::{
-    PaymentPayload, SignedCommand, SignedCommandPayload, SignedCommandPayloadBody,
-    SignedCommandPayloadCommon, UserCommand,
+    PaymentPayload, SignedCommandPayload, SignedCommandPayloadBody, SignedCommandPayloadCommon,
 };
 
 use super::SignedCommandMemo;
@@ -25,14 +24,13 @@ pub struct SignedTransferCommandBuilder {
     memo: SignedCommandMemo,
     fee_payer_pk: CompressedPubKey,
     valid_until: GlobalSlotNumber,
-    network: NetworkId,
 }
 
 impl SignedTransferCommandBuilder {
     /// All required fields must be defined initially
     pub fn new(
-        to: CompressedPubKey,
         from: CompressedPubKey,
+        to: CompressedPubKey,
         amount: Amount,
         fee: Amount,
         nonce: AccountNonce,
@@ -48,7 +46,6 @@ impl SignedTransferCommandBuilder {
             fee_payer_pk: from,
             memo: SignedCommandMemo::default(),
             valid_until: GlobalSlotNumber::MAX,
-            network: NetworkId::TESTNET,
         }
     }
 
@@ -78,14 +75,17 @@ impl SignedTransferCommandBuilder {
         Self { memo, ..self }
     }
 
-    /// Set the network this command can work on (mainnet or test)
-    pub fn network(self, network: NetworkId) -> Self {
-        Self { network, ..self }
+    /// Set the global slot which this command is valid until
+    pub fn valid_until(self, valid_until: GlobalSlotNumber) -> Self {
+        Self {
+            valid_until,
+            ..self
+        }
     }
 
     /// Sign the transaction and produce a UserCommand with the signature fields filled
-    pub fn sign_and_build(self, keypair: Keypair) -> UserCommand {
-        let payload = SignedCommandPayload {
+    pub fn build(self) -> SignedCommandPayload {
+        SignedCommandPayload {
             common: SignedCommandPayloadCommon {
                 fee: self.fee,
                 fee_token: self.fee_token,
@@ -100,17 +100,6 @@ impl SignedTransferCommandBuilder {
                 source_pk: self.from,
                 token_id: self.transfer_token,
             }),
-        };
-
-        // This should change to create_kimchi after fork
-        let mut ctx =
-            proof_systems::mina_signer::create_legacy::<SignedCommandPayload>(self.network);
-        let signature = ctx.sign(&keypair, &payload);
-
-        UserCommand::SignedCommand(SignedCommand {
-            payload,
-            signer: keypair.public.into_compressed(),
-            signature,
-        })
+        }
     }
 }
